src/main/java/danilov/roman/myjava/service/UserService.java
package danilov.roman.myjava.service;

//import danilov.roman.myjava.model.Privilege;
import danilov.roman.myjava.model.Role;
import danilov.roman.myjava.model.User;
import danilov.roman.myjava.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.GrantedAuthority;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.*;

@Service
public class UserService implements UserDetailsService {

    private final UserRepository userRepository;
    private BCryptPasswordEncoder bCryptPasswordEncoder;

    public UserService(UserRepository userRepository, BCryptPasswordEncoder bCryptPasswordEncoder) {
        this.userRepository = userRepository;
        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
    }

    /**
     * Добавление нового пользователя
     *
     * @param user User
     * @return boolean false - если пользователь уже существует, true - если успешно добавился.
     */
    public boolean addUser(User user) {
        User userFromBd = userRepository.findByEmail(user.getEmail());
        if (userFromBd != null) {
            return false;
        }
        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
        userRepository.save(user);
        return true;
    }

    /**
     * Возращает найденного по ID пользователя, иначе null
     *
     * @param id Integer ID пользователя
     * @return User/null
     */
    public User getUserById(int id) {
        return userRepository.findById(id).orElse(null);
    }

    /**
     * Возвращает список всех пользователей
     *
     * @return List<User>
     */
    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    /**
     * Возращает найденного по email пользователя
     *
     * @param email String
     * @return User
     */
    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        User user = userRepository.findByEmail(email);

        if (user == null) {
            throw new UsernameNotFoundException("User not found");
        }
        return user;
    }
}
